My portfolio that I'll submit to the Robotics Team
There are 2 BMI Calculators, one where you input values based on the units of the Imperial system and another where you input values based on units of the metric system. The chemistry program encompasses a variety of functions that answers chemistry questions on different topics. Chem-TableF is uncompleted and is supposed to compliment the Chemistry program.